Fix GST HST threshold progress type

+++ lib/marketplace/marketplace_canada_gst_hst_threshold_panel.dart
@@ -118,7 +118,7 @@ class _MarketplaceCanadaGstHstThresholdPanelState
     final level = '${assessment['level'] ?? 'not_assessed'}';
     final progress = thresholdMinor <= 0
         ? 0.0
-        : (governingMinor / thresholdMinor).clamp(0.0, 1.0);
+        : (governingMinor / thresholdMinor).clamp(0.0, 1.0).toDouble();
     final state = _visualState(level, hasAssessment);
 
     return Card(
